Two-tier approach to overcome GitHub permission restriction
Using a two-tier approach to overcome GitHub permission restriction
on pull requested-based runs.

The caller-corevalidation.yml is triggered by PRs (with limited
permission if forked), it collects PR number to be passed to the
called corevalidation.yml (this run has  full permissions, e.g.
to assume-role and consume secrets).

The corevalidation workflow will feedback on its status to the PR
with a "CoreValidation" check name.

Caveat: The corevalidation.yml code is, by GH design, run from
the base branch, not from the PR. So, changes on integration-test.yml
file only take effect when merged to the base branch (e.g. main)